+++ src/interface/navigation.rs
@@ -14,7 +14,7 @@ impl Events for MobileNavigator {}
 
 impl MobileNavigator {
     pub fn new(ctx: &mut Context) -> Self {
-        MobileNavigator(Row(48, Offset::Center, Size::Fit, Padding(0, 8, 0, 8)), vec![
+        MobileNavigator(Row(48.0, Offset::Center, Size::Fit, Padding(0.0, 8.0, 0.0, 8.0)), vec![
             IconButton::tab_nav(ctx, "wallet", true, |_ctx: &mut Context| println!("Bitcoin")),
             IconButton::tab_nav(ctx, "messages", false, |_ctx: &mut Context| println!("Messaging")),
             IconButton::tab_nav(ctx, "door", false, |_ctx: &mut Context| println!("Rooms")),
@@ -35,11 +35,11 @@ impl DesktopNavigator {
         let messages = Button::navigation(ctx, "messages", "Messages", false, |_ctx: &mut Context| println!("Messaging"));
         let rooms = Button::navigation(ctx, "door", "Rooms", false, |_ctx: &mut Context| println!("Rooms"));
         DesktopNavigator(
-            Column(32, Offset::Center, Size::Fill(100, 200), Padding(16, 32, 16, 32)),
-            Brand::new(wordmark, (80, 44)),
+            Column(32.0, Offset::Center, Size::Fill(100.0, 200.0), Padding(16.0, 32.0, 16.0, 32.0)),
+            Brand::new(wordmark, (80.0, 44.0)),
             ButtonColumn::new(vec![bitcoin, messages, rooms]),
             Bin (
-                Stack(Offset::Center, Offset::Center, Size::Fill(100, 200), Size::Fill(100, u32::MAX),  Padding::default()), 
+                Stack(Offset::Center, Offset::Center, Size::Fill(100.0, 200.0), Size::Fill(100.0, f32::MAX),  Padding::default()), 
                 Rectangle::new(color)
             ),
             Button::photo(ctx, "My Profile", AvatarContent::Icon("profile", AvatarIconStyle::Secondary), false, |_ctx: &mut Context| println!("Profile"))
@@ -55,7 +55,7 @@ impl Events for Header {}
 impl Header {
     pub fn home(ctx: &mut Context, title: &'static str) -> Self {
         Header(
-            Row(16, Offset::Center, Size::Fit, Padding(24, 16, 24, 16)),
+            Row(16.0, Offset::Center, Size::Fit, Padding(24.0, 16.0, 24.0, 16.0)),
             HeaderIcon::new(None), 
             HeaderContent::home(ctx, title),
             HeaderIcon::new(None)
@@ -69,7 +69,7 @@ impl Header {
         right: Option<IconButton>
     ) -> Self {
         Header(
-            Row(16, Offset::Center, Size::Fit, Padding(24, 16, 24, 16)),
+            Row(16.0, Offset::Center, Size::Fit, Padding(24.0, 16.0, 24.0, 16.0)),
             HeaderIcon::new(left), 
             HeaderContent::stack(ctx, title), 
             HeaderIcon::new(right)
@@ -83,7 +83,7 @@ impl Header {
         avatars: Vec<AvatarContent>,
     ) -> Self {
         Header(
-            Row(16, Offset::Center, Size::Fit, Padding(24, 16, 24, 16)),
+            Row(16.0, Offset::Center, Size::Fit, Padding(24.0, 16.0, 24.0, 16.0)),
             HeaderIcon::new(left), 
             HeaderContent::chat(ctx, avatars), 
             HeaderIcon::new(right)
@@ -98,9 +98,9 @@ impl Events for HeaderContent {}
 impl HeaderContent {
     pub fn home(ctx: &mut Context, title: &'static str) -> Self {
         let text_size = ctx.get::<PelicanUI>().theme.fonts.size.h3;
-        let width = Size::custom(move |widths: Vec<(u32, u32)>|(widths[0].0, u32::MAX));
+        let width = Size::custom(move |widths: Vec<(f32, f32)>|(widths[0].0, f32::MAX));
         HeaderContent(
-            Column(10, Offset::Center, width, Padding::default()), 
+            Column(10.0, Offset::Center, width, Padding::default()), 
             Text::new(ctx, title, TextStyle::Heading, text_size),
             None,
         )
@@ -108,9 +108,9 @@ impl HeaderContent {
 
     pub fn stack(ctx: &mut Context, title: &'static str) -> Self {
         let text_size = ctx.get::<PelicanUI>().theme.fonts.size.h4;
-        let width = Size::custom(move |widths: Vec<(u32, u32)>|(widths[0].0, u32::MAX));
+        let width = Size::custom(move |widths: Vec<(f32, f32)>|(widths[0].0, f32::MAX));
         HeaderContent(
-            Column(10, Offset::Center, width, Padding::default()),  
+            Column(10.0, Offset::Center, width, Padding::default()),  
             Text::new(ctx, title, TextStyle::Heading, text_size),
             None,
         )
@@ -119,9 +119,9 @@ impl HeaderContent {
     pub fn chat(ctx: &mut Context, avatars: Vec<AvatarContent>) -> Self {
         let text_size = ctx.get::<PelicanUI>().theme.fonts.size.h5;
         let title = if avatars.len() > 1 {"Ella Couch"} else {"Group Message"};
-        let width = Size::custom(move |widths: Vec<(u32, u32)>|(widths[0].0, u32::MAX));
+        let width = Size::custom(move |widths: Vec<(f32, f32)>|(widths[0].0, f32::MAX));
         HeaderContent(
-            Column(10, Offset::Center, width, Padding::default()), 
+            Column(10.0, Offset::Center, width, Padding::default()), 
             Text::new(ctx, title, TextStyle::Heading, text_size),
             Some(AvatarRow::new(ctx, avatars)),
         )
@@ -135,7 +135,7 @@ impl Events for HeaderIcon {}
 impl HeaderIcon {
     pub fn new(icon: Option<IconButton>) -> Self {
         HeaderIcon(
-            Stack(Offset::Center, Offset::Center, Size::Static(48), Size::Static(48), Padding::default()),
+            Stack(Offset::Center, Offset::Center, Size::Static(48.0), Size::Static(48.0), Padding::default()),
             icon
         )
     }
